validate order api

+++ src/main/java/com/cg/api/OrderAPI.java
@@ -8,12 +8,18 @@ import com.cg.model.dto.order.OrderResDTO;
 import com.cg.model.dto.orderDetail.OrderDetailByTableResDTO;
 import com.cg.service.order.IOrderService;
 import com.cg.service.orderDetail.IOrderDetailService;
+import com.cg.utils.AppUtils;
+import com.cg.utils.ValidateUtils;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
+import org.springframework.validation.BindingResult;
 import org.springframework.web.bind.annotation.*;
 
+import java.util.HashMap;
 import java.util.List;
+import java.util.Map;
+import java.util.Optional;
 
 @RestController
 @RequestMapping("/api/orders")
@@ -24,24 +30,40 @@ public class OrderAPI {
     @Autowired
     private IOrderDetailService orderDetailService;
 
+    @Autowired
+    private ValidateUtils validateUtils;
 
+    @Autowired
+    private AppUtils appUtils;
 
 
     @GetMapping("/table/{tableId}")
-    public ResponseEntity<?> getOrderByTableId(@PathVariable Long tableId) {
+    public ResponseEntity<?> getOrderByTableId(@PathVariable("tableId") String tableIdStr) {
+        if (!validateUtils.isNumberValid(tableIdStr)) {
+            throw new DataInputException("Mã bàn không hợp lệ");
+        }
+        Long tableId = Long.valueOf(tableIdStr);
 
         Order order = orderService.findByTableId(tableId).orElseThrow(() -> {
-           throw new DataInputException("Bàn không tồn tại");
+            throw new DataInputException("Bàn không tồn tại");
         });
 
         List<OrderDetailByTableResDTO> getOrderDetailByTableResDTO = orderDetailService.getOrderDetailByTableResDTO(order.getId());
 
+        if (getOrderDetailByTableResDTO.size() == 0) {
+            return new ResponseEntity<>(HttpStatus.NO_CONTENT);
+        }
         return new ResponseEntity<>(getOrderDetailByTableResDTO, HttpStatus.OK);
     }
 
     // tạo mới order
     @PostMapping
-    public ResponseEntity<?> saveOrder(@RequestBody OrderReqDTO orderReqDTO){
+    public ResponseEntity<?> saveOrder(@RequestBody OrderReqDTO orderReqDTO, BindingResult bindingResult) {
+
+        new OrderReqDTO().validate(orderReqDTO, bindingResult);
+        if (bindingResult.hasFieldErrors()) {
+            return appUtils.mapErrorToResponse(bindingResult);
+        }
 
 
         OrderResDTO orderResDTO = orderService.createOrder(orderReqDTO);
@@ -50,29 +72,42 @@ public class OrderAPI {
 
     // thêm order vào order củ
     @PostMapping("/{id}/order-details")
-    public ResponseEntity<?> createOrderDetail(@RequestBody OrderReqDTO orderReqDTO, @PathVariable Long id){
+    public ResponseEntity<?> createOrderDetail(@RequestBody OrderReqDTO orderReqDTO, @PathVariable("id") String idstr, BindingResult bindingResult) {
+        if (!validateUtils.isNumberValid(idstr)) {
+            throw new DataInputException("Mã đặt hàng không hợp lệ");
+        }
+
+        new OrderReqDTO().validate(orderReqDTO, bindingResult);
+        if (bindingResult.hasFieldErrors()) {
+            return appUtils.mapErrorToResponse(bindingResult);
+        }
+        Long id = Long.valueOf(idstr);
 
         OrderResDTO orderResDTO = orderService.updateOrderDetail(orderReqDTO, id);
-        return new ResponseEntity<>(orderResDTO, HttpStatus.CREATED);
+        return new ResponseEntity<>(orderResDTO, HttpStatus.OK);
     }
-
 
 
     // xóa order
     @DeleteMapping("/{orderId}/order-details/{orderDetailId}")
     public ResponseEntity<?> deleteOrder(@PathVariable Long orderId, @PathVariable Long orderDetailId) {
         Order order = orderService.findById(orderId).orElseThrow(() -> {
-           throw new DataInputException("sản phẩm không tồn tại");
+            throw new DataInputException("Mã đặt hàng không tồn tại");
         });
 
-        OrderDetail orderDetail = orderDetailService.findById(orderDetailId).orElseThrow(() -> {
-           throw new DataInputException("Mã sản phẩm không tồn tại");
-        });
+//        OrderDetail orderDetail = orderDetailService.findById(orderDetailId).orElseThrow(() -> {
+//           throw new DataInputException("Mã sản phẩm không tồn tại");
+//        });
+        Optional<OrderDetail> optionalOrderDetail = orderDetailService.findById(orderDetailId);
+        if (optionalOrderDetail.isEmpty()) {
+            Map<String, String> data = new HashMap<>();
+            data.put("message", "Mã sản phẩm không tồn tại");
+            return new ResponseEntity<>(data, HttpStatus.BAD_REQUEST);
+        }
 
 
         orderService.deleteByIdOrder(orderId, orderDetailId);
 
-        order = orderService.findById(orderId).get();
         OrderResDTO orderResDTO = order.toOrderResDTO();
         return new ResponseEntity<>(orderResDTO, HttpStatus.OK);
     }
